requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
slowapi==0.1.9
beautifulsoup4==4.12.3
httpx>=0.27.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
asyncpg>=0.29.0
langchain-core>=0.2.27
langchain-groq>=0.1.5